chore: breaking up StarSearch into components
